VintageStory, packaging

+++ script/vintagestory.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
 
-sudo pacman -S flatpak
-flatpak --user install https://flatpak.vintagestory.at/VintageStory.flatpakref
+if gum confirm "Install VintageStory? (Semi-closed, requires paid account)"; then
+	sudo pacman -S flatpak
+	flatpak --user install https://flatpak.vintagestory.at/VintageStory.flatpakref
+fi
